Move crc16 to helpers (#3780)
