+++ esphome/components/am2320/am2320.cpp
@@ -4,32 +4,14 @@
 //  - Arduino - AM2320: https://github.com/EngDial/AM2320/blob/master/src/AM2320.cpp
 
 #include "am2320.h"
-#include "esphome/core/log.h"
 #include "esphome/core/hal.h"
+#include "esphome/core/helpers.h"
+#include "esphome/core/log.h"
 
 namespace esphome {
 namespace am2320 {
 
 static const char *const TAG = "am2320";
-
-// ---=== Calc CRC16 ===---
-uint16_t crc_16(uint8_t *ptr, uint8_t length) {
-  uint16_t crc = 0xFFFF;
-  uint8_t i;
-  //------------------------------
-  while (length--) {
-    crc ^= *ptr++;
-    for (i = 0; i < 8; i++) {
-      if ((crc & 0x01) != 0) {
-        crc >>= 1;
-        crc ^= 0xA001;
-      } else {
-        crc >>= 1;
-      }
-    }
-  }
-  return crc;
-}
 
 void AM2320Component::update() {
   uint8_t data[8];
@@ -98,7 +80,7 @@ bool AM2320Component::read_data_(uint8_t *data) {
   checksum = data[7] << 8;
   checksum += data[6];
 
-  if (crc_16(data, 6) != checksum) {
+  if (crc16(data, 6) != checksum) {
     ESP_LOGW(TAG, "AM2320 Checksum invalid!");
     return false;
   }
