+++ esphome/core/helpers.h
@@ -149,6 +149,9 @@ template<typename T, typename U> T remap(U value, U min, U max, T min_out, T max
 /// Calculate a CRC-8 checksum of \p data with size \p len.
 uint8_t crc8(uint8_t *data, uint8_t len);
 
+/// Calculate a CRC-16 checksum of \p data with size \p len.
+uint16_t crc16(const uint8_t *data, uint8_t len);
+
 /// Calculate a FNV-1 hash of \p str.
 uint32_t fnv1_hash(const std::string &str);
